+++ docs/source/Changelog.rst
@@ -95,6 +95,7 @@ Fixed
 -  BoM
 
    -  The field name ``Reference`` was accepted, but didn’t work
+   -  Missing footprint information for sub-units (#741)
 
 -  Do Not Fit footprint crosses: